news-summarizer/requirements.txt
requests
newspaper3k
langchain
python-dotenv
deeplake
lxml_html_clean
langchain_google_genai
